requirements.txt
aiogram>=3.0.0
aiosqlite
aiohttp
aiohttp_cors
python-dotenv
